break up some monolithic chunks in r session wrappers (WIP)
