+++ src/cpp/r/CMakeLists.txt
@@ -37,10 +37,12 @@ set (R_SOURCE_FILES
    session/RConsoleActions.cpp
    session/RConsoleHistory.cpp
    session/RDiscovery.cpp
+   session/RQuit.cpp
    session/RRestartContext.cpp
    session/RSearchPath.cpp
    session/RSessionState.cpp
    session/RSession.cpp
+   session/RStdCallbacks.cpp
    session/graphics/RGraphicsDevice.cpp
    session/graphics/RGraphicsErrorCategory.cpp
    session/graphics/RGraphicsPlot.cpp
