+++ src/cpp/r/session/RSession.cpp
@@ -1,7 +1,7 @@
 /*
  * RSession.cpp
  *
- * Copyright (C) 2009-12 by RStudio, Inc.
+ * Copyright (C) 2009-18 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -18,7 +18,6 @@
 
 #include <iostream>
 
-#include <boost/regex.hpp>
 #include <boost/algorithm/string/predicate.hpp>
 #include <boost/algorithm/string/replace.hpp>
 #include <boost/algorithm/string/trim.hpp>
@@ -31,7 +30,6 @@
 #include <core/system/Environment.hpp>
 #include <core/FileSerializer.hpp>
 #include <core/FileUtils.hpp>
-#include <core/RegexUtils.hpp>
 #include <core/http/Util.hpp>
 
 #include <r/RExec.hpp>
@@ -43,15 +41,16 @@
 #include <r/RInterface.hpp>
 #include <r/RFunctionHook.hpp>
 #include <r/session/RSessionState.hpp>
-#include <r/session/RConsoleActions.hpp>
 #include <r/session/RConsoleHistory.hpp>
 #include <r/session/RClientState.hpp>
 #include <r/session/RGraphics.hpp>
 #include <r/session/RDiscovery.hpp>
 
 #include "RClientMetrics.hpp"
-#include "RRestartContext.hpp"
 #include "REmbedded.hpp"
+#include "RQuit.hpp"
+#include "RRestartContext.hpp"
+#include "RStdCallbacks.hpp"
 
 #include "graphics/RGraphicsDevDesc.hpp"
 #include "graphics/RGraphicsUtils.hpp"
@@ -82,19 +81,13 @@ __declspec(dllimport) SA_TYPE SaveAction;
 // constants for graphics scratch subdirectory
 #define kGraphicsPath "graphics"
 
-using namespace rstudio::core ;
+using namespace rstudio::core;
 
 namespace rstudio {
 namespace r {
 namespace session {
 
 namespace {
-
-// is this R 3.0 or greator
-bool s_isR3 = false;
-
-// is this R 3.3 or greator
-bool s_isR3_3 = false;
 
 // options
 ROptions s_options;
@@ -115,14 +108,11 @@ FilePath s_suspendedSessionPath ;
 // latent deserialization actions are taking place
 boost::function<void()> s_deferredDeserializationAction;
    
-// have we completed our one-time initialization yet?
-bool s_initialized = false;
-
 // are in the middle of servicing a suspend request?
 bool s_suspended = false;
 
-// temporarily suppress output
-bool s_suppressOutput = false;
+// script to run, if any
+std::string s_runScript;
 
 FilePath rHistoryFilePath()
 {
@@ -400,482 +390,12 @@ void restoreSession(const FilePath& suspendedSessionPath,
    }
 }
 
-// one-time per session initialization
-Error initialize()
-{
-   // ensure that the utils package is loaded (it might not be loaded
-   // if R is attempting to recover from a library loading error which
-   // occurs during .Rprofile)
-   Error libError = r::exec::RFunction("library", "utils").call();
-   if (libError)
-      LOG_ERROR(libError);
-
-   // check whether this is R 3.3 or greater
-   Error r33Error = r::exec::evaluateString("getRversion() >= '3.3.0'", &s_isR3_3);
-   if (r33Error)
-      LOG_ERROR(r33Error);
-
-   if (s_isR3_3)
-   {
-      s_isR3 = true;
-   }
-   else
-   {
-      // check whether this is R 3.0 or greater
-      Error r3Error = r::exec::evaluateString("getRversion() >= '3.0.0'", &s_isR3);
-      if (r3Error)
-         LOG_ERROR(r3Error);
-   }
-
-   // initialize console history capacity
-   r::session::consoleHistory().setCapacityFromRHistsize();
-
-   // install R tools
-   FilePath toolsFilePath = s_options.rSourcePath.complete("Tools.R");
-   Error error = r::sourceManager().sourceTools(toolsFilePath);
-   if (error)
-      return error ;
-
-   // install RStudio API
-   FilePath apiFilePath = s_options.rSourcePath.complete("Api.R");
-   error = r::sourceManager().sourceTools(apiFilePath);
-   if (error)
-      return error;
-
-   // initialize graphics device -- use a stable directory for server mode
-   // and temp directory for desktop mode (so that we can support multiple
-   // concurrent processes using the same project)
-   FilePath graphicsPath;
-   if (s_options.serverMode)
-   {
-      std::string path = kGraphicsPath;
-      if (utils::isR3())
-         path += "-r3";
-      graphicsPath = s_options.sessionScratchPath.complete(path);
-   }
-   else
-   {
-      graphicsPath = r::session::utils::tempDir().complete(
-                              "rs-graphics-" + core::system::generateUuid());
-   }
-
-   error = graphics::device::initialize(graphicsPath,
-                                        s_callbacks.locator);
-   if (error) 
-      return error;
-   
-   // restore client state
-   session::clientState().restore(s_clientStatePath,
-                                  s_projectClientStatePath);
-      
-   // restore suspended session if we have one
-   bool wasResumed = false;
-   
-   // first check for a pending restart
-   if (restartContext().hasSessionState())
-   {
-      // restore session
-      std::string errorMessages ;
-      restoreSession(restartContext().sessionStatePath(), &errorMessages);
-
-      // show any error messages
-      if (!errorMessages.empty())
-         REprintf(errorMessages.c_str());
-
-      // note we were resumed
-      wasResumed = true;
-   }
-   else if (s_suspendedSessionPath.exists())
-   {  
-      // restore session
-      std::string errorMessages ;
-      restoreSession(s_suspendedSessionPath, &errorMessages);
-      
-      // show any error messages
-      if (!errorMessages.empty())
-         REprintf(errorMessages.c_str());
-
-      // note we were resumed
-      wasResumed = true;
-   }  
-   // new session
-   else
-   {  
-      // restore console history
-      FilePath historyPath = rHistoryFilePath();
-      error = consoleHistory().loadFromFile(historyPath, false);
-      if (error)
-         reportHistoryAccessError("read history from", historyPath, error);
-
-      // defer loading of global environment
-      s_deferredDeserializationAction = deferredRestoreNewSession;
-   }
-   
-   // initialize client
-   RInitInfo rInitInfo(wasResumed);
-   error = s_callbacks.init(rInitInfo);
-   if (error)
-      return error;
-
-   // call resume hook if we were resumed
-   if (wasResumed)
-      s_callbacks.resumed();
-   
-   // now that all initialization code has had a chance to run we 
-   // can register all external routines which were added to r::routines
-   // during the init sequence
-   r::routines::registerAll();
-   
-   // set default repository if requested
-   if (!s_options.rCRANRepos.empty())
-   {
-      error = r::exec::RFunction(".rs.setCRANReposAtStartup",
-                                 s_options.rCRANRepos).call();
-      if (error)
-         return error;
-   }
-
-   // initialize profile resources
-   error = r::exec::RFunction(".rs.profileResources").call();
-   if (error)
-      return error;
-
-   // complete embedded r initialization
-   error = r::session::completeEmbeddedRInitialization(s_options.useInternet2);
-   if (error)
-      return error;
-
-   // set global R options
-   FilePath optionsFilePath = s_options.rSourcePath.complete("Options.R");
-   error = r::sourceManager().sourceLocal(optionsFilePath);
-   if (error)
-      return error;
-
-   // server specific R options options
-   if (s_options.serverMode)
-   {
-#ifndef __APPLE__
-      FilePath serverOptionsFilePath =  s_options.rSourcePath.complete(
-                                                         "ServerOptions.R");
-      return r::sourceManager().sourceLocal(serverOptionsFilePath);
-#else
-      return Success();
-#endif
-   }
-   else
-   {
-      return Success();
-   }
-}
-
-void rSuicide(const std::string& msg)
-{
-   // log abort message if we are in desktop mode
-   if (!s_options.serverMode)
-   {
-      FilePath abendLogPath = s_options.logPath.complete(
-                                                 "rsession_abort_msg.log");
-      Error error = core::writeStringToFile(abendLogPath, msg);
-      if (error)
-         LOG_ERROR(error);
-   }
-
-
-   R_Suicide(msg.c_str());
-}
-
-void rSuicide(const Error& error)
-{
-   // provide error message if the error was unexpected
-   std::string msg;
-   if (!error.expected())
-      msg = core::log::errorAsLogEntry(error);
-
-   rSuicide(msg);
-}
-
-// forward declare win32 quit handler and provide string based quit
-// handler that parses the quit command from the console
-#ifdef _WIN32
-bool win32Quit(const std::string& saveAction,
-               int status,
-               bool runLast,
-               std::string* pErrMsg);
-
-bool win32Quit(const std::string& command, std::string* pErrMsg)
-{
-   // default values
-   std::string saveAction = "default";
-   double status = 0;
-   bool runLast = true;
-
-   // parse quit arguments
-   SEXP argsSEXP;
-   r::sexp::Protect rProtect;
-   Error error = r::exec::RFunction(".rs.parseQuitArguments", command).call(
-                                                                     &argsSEXP,
-                                                                     &rProtect);
-   if (!error)
-   {
-      error = r::sexp::getNamedListElement(argsSEXP,
-                                           "save",
-                                           &saveAction,
-                                           saveAction);
-      if (error)
-         LOG_ERROR(error);
-
-      error = r::sexp::getNamedListElement(argsSEXP,
-                                           "status",
-                                           &status,
-                                           status);
-      if (error)
-         LOG_ERROR(error);
-
-      error = r::sexp::getNamedListElement(argsSEXP,
-                                           "runLast",
-                                           &runLast,
-                                           runLast);
-      if (error)
-         LOG_ERROR(error);
-   }
-   else
-   {
-      *pErrMsg = r::endUserErrorMessage(error);
-      return false;
-   }
-
-   return win32Quit(saveAction, static_cast<int>(status), runLast, pErrMsg);
-}
-
-#endif
-
-bool consoleInputHook(const std::string& prompt,
-                      const std::string& input)
-{
-   // only check for quit when we're at the default prompt
-   if (!r::session::utils::isDefaultPrompt(prompt))
-      return true;
-
-   // check for user quit invocation
-    boost::regex re("^\\s*(q|quit)\\s*\\(.*$");
-    boost::smatch match;
-    if (regex_utils::match(input, match, re))
-   {
-      if (!s_callbacks.handleUnsavedChanges())
-      {
-         REprintf("User cancelled quit operation\n");
-         return false;
-      }
-
-      // on win32 we will actually assume responsibility for the
-      // quit function entirely (so we can call our internal cleanup
-      // handler code)
-#ifdef _WIN32
-      std::string quitErr;
-      bool didQuit = win32Quit(input, &quitErr);
-      if (!didQuit)
-         REprintf((quitErr + "\n").c_str());
-
-      // always return false (since we take over the command fully)
-      return false;
-#else
-      return true;
-#endif
-   }
-   else
-   {
-      return true;
-   }
-}
-
 bool isInjectedBrowserCommand(const std::string& cmd)
 {
    return browserContextActive() &&
           (cmd == "c" || cmd == "Q" || cmd == "n" || cmd == "s" || cmd == "f");
 }
 
-
-int RReadConsole (const char *pmt,
-                  CONSOLE_BUFFER_CHAR* buf,
-                  int buflen,
-                  int hist)
-{
-   try
-   {
-      // capture the prompt for later manipulation
-      std::string prompt(pmt);
-
-      // invoke one time initialization
-      if (!s_initialized)
-      {
-         // ignore interrupts which occur during initialization. any
-         // interrupt will cause the initialization to fail which will
-         // then require the user to start over from the beginning. if the
-         // user has to wait in any case we might as well help them out by
-         // never having to start from scratch
-         r::exec::IgnoreInterruptsScope ignoreInterrupts;
-         
-         // attempt to initialize 
-         Error initError;
-         Error error = r::exec::executeSafely<Error>(initialize, &initError);
-         if (error || initError)
-         {
-            if (initError)
-               error = initError;
-
-            // log the error if it was unexpected
-            if (!error.expected())
-               LOG_ERROR(error);
-            
-            // terminate the session (use suicide so that no special
-            // termination code runs -- i.e. call to setAbnormalEnd(false)
-            // or call to client::quitSession)
-            rSuicide(error);
-         }
-         
-         // reset the prompt to whatever the default is after we've
-         // fully initialized (and restored suspended options)
-         prompt = r::options::getOption<std::string>("prompt");
-
-         // ensure only one initialization 
-         s_initialized = true;
-      }
-
-      std::string promptString(prompt);
-      promptString = util::rconsole2utf8(promptString);
-
-      // get the next input
-      bool addToHistory = (hist == 1);
-      RConsoleInput consoleInput("");
-      if ( s_callbacks.consoleRead(promptString, addToHistory, &consoleInput) )
-      {
-         // add prompt to console actions (we do this after consoleRead
-         // completes so that we don't send both a console prompt event
-         // AND include the same prompt in the actions history)
-         consoleActions().add(kConsoleActionPrompt, prompt);
-
-         if (consoleInput.cancel)
-         {
-            // notify of interrupt
-            consoleActions().notifyInterrupt();
-
-            // escape out using exception so that we can allow normal
-            // c++ stack unwinding to occur before jumping
-            throw r::exec::InterruptException();
-         }
-         else
-         {
-            // determine the input to return to R
-            std::string rInput = consoleInput.text;
-
-            // refresh source if necessary (no-op in production)
-            r::sourceManager().reloadIfNecessary();
-            
-            // ensure that our input fits within the buffer
-            std::string::size_type maxLen = buflen - 2; // for \n\0
-            rInput = string_utils::utf8ToSystem(rInput, true);
-            if (rInput.length() > maxLen)
-               rInput.resize(maxLen);
-            std::string::size_type inputLen = rInput.length();
-            
-            // add to console actions and history (if requested). note that
-            // we add the user's input rather than any tranformed input we
-            // created as a result of a shell escape
-            consoleActions().add(kConsoleActionInput, consoleInput.text);
-            if (addToHistory && !isInjectedBrowserCommand(consoleInput.text))
-               consoleHistory().add(consoleInput.text);
-
-            // call console input hook and interrupt if the hook tells us to
-            if (!consoleInputHook(prompt, consoleInput.text))
-               throw r::exec::InterruptException();
-
-            // copy to buffer and add terminators
-            rInput.copy( (char*)buf, maxLen);
-            buf[inputLen] = '\n';
-            buf[inputLen+1] = '\0';
-         }
-
-         return 1 ;
-      }
-      else
-      {
-         return 0; // terminate
-      }
-   }
-   catch(r::exec::InterruptException&)
-   {
-      // set interrupts pending
-      r::exec::setInterruptsPending(true);
-
-      // only issue an interrupt when not on Windows -- let the regular
-      // event loop handle interrupts there. note that this will longjmp
-#ifndef _WIN32
-      r::exec::checkUserInterrupt();
-#endif
-
-      // return success
-      return 1;
-   }
-   catch(const std::exception& e)
-   {
-      std::string msg = std::string("Unexpected exception: ") + e.what();
-      LOG_ERROR_MESSAGE(msg);
-      rSuicide(msg);
-   }
-   catch(...)
-   {
-      std::string msg = "Unknown exception";
-      LOG_ERROR_MESSAGE(msg);
-      rSuicide(msg);
-   }
-      
-   return 0 ; // keep compiler happy
-}
-   
-void RWriteConsoleEx (const char *buf, int buflen, int otype)
-{
-   try
-   {
-      if (!s_suppressOutput)
-      {
-         // get output
-         std::string output = std::string(buf,buflen);
-         output = util::rconsole2utf8(output);
-         
-         // add to console actions
-         int type = otype == 1 ? kConsoleActionOutputError :
-                                 kConsoleActionOutput;
-         consoleActions().add(type, output);
-         
-         // write
-         s_callbacks.consoleWrite(output, otype) ;
-      }
-   }
-   CATCH_UNEXPECTED_EXCEPTION
-}
-
-void RShowMessage(const char* msg)
-{
-   try 
-   {
-      s_callbacks.showMessage(msg) ;
-   }
-   CATCH_UNEXPECTED_EXCEPTION
-}
-
-   
-// NOTE: Win32 doesn't receive this callback
-int REditFile(const char* file)
-{
-   try 
-   {
-      return s_callbacks.editFile(r::util::fixPath(file));
-   }
-   CATCH_UNEXPECTED_EXCEPTION
-   
-   // error if we got this far
-   return 1 ;
-}
 
 SEXP rs_editFile(SEXP fileSEXP)
 {
@@ -921,83 +441,6 @@ SEXP rs_showFile(SEXP titleSEXP, SEXP fileSEXP, SEXP delSEXP)
    return R_NilValue;
 }
 
-// NOTE: Win32 doesn't receive this callback
-int RShowFiles (int nfile, 
-                const char **file, 
-                const char **headers, 
-                const char *wtitle, 
-                Rboolean del, 
-                const char *pager)
-{
-   try 
-   {
-      for (int i=0; i<nfile; i++)
-      {
-         // determine file path and title
-         std::string fixedPath = r::util::fixPath(file[i]);
-         FilePath filePath = utils::safeCurrentPath().complete(fixedPath);
-         if (filePath.exists())
-         {
-            std::string title(headers[i]);
-            if (title.empty())
-               title = wtitle;
-         
-            // show file
-            s_callbacks.showFile(title, filePath, del);
-         }
-         else
-         {
-            throw r::exec::RErrorException(
-                               "File " + fixedPath + " does not exist.");
-         }
-      }
-   }
-   catch(r::exec::RErrorException& e)
-   {
-      r::exec::error(e.message());
-   }
-
-   CATCH_UNEXPECTED_EXCEPTION
-   
-   // NOTE: the documentation doesn't indicate what to return and do_fileshow
-   // in platform.c doesn't check the return value s
-   return 0;
-}
-
-// NOTE: Win32 doesn't receive this callback
-int RChooseFile (int newFile, char *buf, int len) 
-{
-   try 
-   {
-      FilePath filePath = s_callbacks.chooseFile(newFile == TRUE);
-      if (!filePath.empty())
-      {
-         // get absolute path
-         std::string absolutePath = filePath.absolutePath();
-         
-         // trunate file if it is too long
-         std::string::size_type maxLen = len - 1; 
-         if (absolutePath.length() > maxLen)
-            absolutePath.resize(maxLen);
-         
-         // copy the file to the buffer
-         absolutePath.copy(buf, maxLen);
-         buf[absolutePath.length()] = '\0';
-         
-         // return the length of the filepath buffer
-         return absolutePath.length();
-      }
-      else
-      {
-         return 0;
-      }
-   }
-   CATCH_UNEXPECTED_EXCEPTION
-
-   // error if we got this far
-   return 0 ;
-}
-         
 // method called from browseUrl
 SEXP rs_browseURL(SEXP urlSEXP)
 {
@@ -1071,93 +514,6 @@ SEXP rs_saveHistory(SEXP sFile)
    return R_NilValue;
 }
 
-void doHistoryFileOperation(SEXP args, 
-                            boost::function<Error(const FilePath&)> fileOp)
-{
-   // validate filename argument
-   SEXP file = CAR(args);
-   if (!sexp::isString(file) || sexp::length(file) < 1)
-      throw r::exec::RErrorException("invalid 'file' argument");
-   
-   // calculate full path to history file
-   FilePath historyFilePath(R_ExpandFileName(Rf_translateChar(STRING_ELT(file, 
-                                                                         0))));
-   // perform operation
-   Error error = fileOp(historyFilePath);
-   if (error)
-      throw r::exec::RErrorException(error.code().message());
-}
-   
-void Rloadhistory(SEXP call, SEXP op, SEXP args, SEXP env)
-{
-   try
-   {
-      doHistoryFileOperation(args, boost::bind(&ConsoleHistory::loadFromFile,
-                                               &consoleHistory(), _1, true));
-      
-      s_callbacks.consoleHistoryReset();
-   }
-   catch(r::exec::RErrorException& e)
-   {
-      r::exec::errorCall(call, e.message());
-   }
-}
-   
-void Rsavehistory(SEXP call, SEXP op, SEXP args, SEXP env)
-{
-   try
-   {
-      doHistoryFileOperation(args, boost::bind(&ConsoleHistory::saveToFile,
-                                               &consoleHistory(), _1));
-   }
-   catch(r::exec::RErrorException& e)
-   {
-      r::exec::errorCall(call, e.message());
-   }
-}
-   
-void Raddhistory(SEXP call, SEXP op, SEXP args, SEXP env)
-{
-   try
-   {
-      // get commands
-      std::vector<std::string> commands ;
-      Error error = sexp::extract(CAR(args), &commands);
-      if (error)
-         throw r::exec::RErrorException(error.code().message());
-      
-      // append them
-      ConsoleHistory& history = consoleHistory();
-      std::for_each(commands.begin(), 
-                    commands.end(),
-                    boost::bind(&ConsoleHistory::add, &history, _1));
-   }
-   catch(r::exec::RErrorException& e)
-   {
-      r::exec::errorCall(call, e.message());
-   }
-}
-
-void RBusy(int which)   
-{
-   try
-   {
-      s_callbacks.busy(which == 1) ;
-   }
-   CATCH_UNEXPECTED_EXCEPTION 
-}
-
-
-// track internal callbacks for delgation
-r::session::InternalCallbacks s_internalCallbacks;
-
-// NOTE: Win32 doesn't receive this callback
-void RSuicide(const char* s)
-{
-   s_callbacks.suicide(s);
-   s_internalCallbacks.suicide(s);
-}
-
 class JumpToTopException
 {
 };
@@ -1208,162 +564,6 @@ SA_TYPE saveAsk()
    return SA_SAVE;
 }
 
-// NOTE: Win32 doesn't receive this function. As a result we only use
-// it in server mode (where it is used to support suspending as well as
-// notifying the browser of quit). In desktop mode we allow standard
-// R_CleanUp processing to take place. There are two important implications
-// of this:
-//
-//   (1) In desktop mode we override do_quit and use it as an indicator
-//       that we should save history and persistent client state (an
-//       alternative would be to more eagerly persist this data)
-//
-//   (2) In desktop mode we don't receive termination oriented events such
-//       as quit, suicide, and cleanup. This means that the desktop process
-//       must simply detect that we have exited and terminate itself. It also
-//       means that cleanup of our http damons, file monitoring, etc. never
-//       occurs (not an issue b/c our process is going away but worth noting)
-//
-void RCleanUp(SA_TYPE saveact, int status, int runLast)
-{
-   // perform cleanup that is coupled to our internal history,
-   // environment-saving, client-state, and graphics implementations
-   // and then delegate to internal R_CleanUp for the remainder
-   // of processing
-   try
-   {
-      // set to default if requested
-      if (saveact == SA_DEFAULT)
-         saveact = SaveAction ;
-      
-      // prompt user to resolve SA_SAVEASK into SA_SAVE or SA_NOSAVE
-      if (saveact == SA_SAVEASK) 
-      {
-         if (imageIsDirty() || !s_options.alwaysSaveHistory())
-            saveact = saveAsk(); // can Rf_jump_to_toplevel()
-         else
-            saveact = SA_NOSAVE; // auto-resolve to no save when not dirty
-      }
-
-      // if the session was quit by the user then run our termination code
-      bool sessionQuitByUser = (saveact != SA_SUICIDE) && !s_suspended ;
-      if (sessionQuitByUser)
-      {
-         // run last if requested (can throw error)
-         if (runLast)
-            R_dot_Last();
-         
-         // save history if we either always save history or saveact == SA_SAVE
-         if (s_options.alwaysSaveHistory() || saveact == SA_SAVE)
-         {
-            FilePath historyPath = rHistoryFilePath();
-            Error error = consoleHistory().saveToFile(historyPath);
-            if (error)
-               reportHistoryAccessError("write history to", historyPath, error);
-         }
-
-         // save environment and history
-         if (saveact == SA_SAVE)
-         {
-            // attempt save if the image is dirty
-            if (imageIsDirty())
-            {
-               // attempt to save global environment. raise error (longjmp 
-               // back to REPL) if there was a problem saving
-               Error error = saveDefaultGlobalEnvironment();
-               if (error)
-                  r::exec::error(r::endUserErrorMessage(error));
-            }
-            
-            // update state
-            saveact = SA_NOSAVE;     // prevent R from saving
-         }
-         
-         // since we've successfully completed the session we can safely
-         // remove any serialized session remaining on disk. note that if 
-         // we do not successfully destroy the session then this would cause
-         // data loss when the previous session is read rather than the
-         // contents of .RData. therefore, we refuse to quit if we can't
-         // successfully destroy the suspended session
-         if (!r::session::state::destroy(s_suspendedSessionPath))
-         {
-            // this will cause us to jump back to the REPL loop
-            r::exec::error("Unable to quit (session cleanup failure)\n");
-         }
-
-         // commit client state
-         saveClientState(ClientStateCommitPersistentOnly);
-
-         // clear display
-         r::session::graphics::display().clear();
-                
-         // notify client that the session has been quit
-         s_callbacks.quit();
-      }
-
-      // allow client to cleanup
-      bool terminatedNormally = saveact != SA_SUICIDE;
-      s_callbacks.cleanup(terminatedNormally);
-      
-      // call internal cleanup (never .runLast because we do it above)
-      // NOTE: may want to replace RCleanUp entirely so that the client
-      // can see any errors which occur during cleanup in the console
-      // (they aren't seen now because the handling of quit obstructs them)
-      s_internalCallbacks.cleanUp(saveact, status, FALSE);
-   }
-   CATCH_UNEXPECTED_EXCEPTION 
-}
-
-// Replace the quit function so we can call our R_CleanUp hook. Note
-// that we need to take special measures to code this safely visa-vi
-// Rf_error long jmps and C++ exceptions. Currently, the RCleanUp
-// function can still long jump if runLast encounters an error. We
-// should re-write the combination of this function and RCleanUp to
-// be fully "error-safe" (not doing this now due to regression risk)
-#ifdef _WIN32
-bool win32Quit(const std::string& saveAction,
-               int status,
-               bool runLast,
-               std::string* pErrMsg)
-{
-   if (r::session::browserContextActive())
-   {
-      *pErrMsg = "unable to quit when browser is active";
-      return false;
-   }
-
-   // determine save action
-   SA_TYPE action = SA_DEFAULT;
-   if (saveAction == "ask")
-      action = SA_SAVEASK;
-   else if (saveAction == "no")
-      action = SA_NOSAVE;
-   else if (saveAction == "yes")
-      action = SA_SAVE;
-   else if (saveAction == "default")
-      action = SA_DEFAULT;
-   else
-   {
-      *pErrMsg = "Unknown save action: " + saveAction;
-      return false;
-   }
-
-   // clean up
-   Error error = r::exec::executeSafely(
-                  boost::bind(&RCleanUp, action, status, runLast));
-   if (error)
-   {
-      *pErrMsg = r::endUserErrorMessage(error);
-      return false;
-   }
-
-   // failsafe in case we don't actually quit as a result of cleanup
-   ::exit(0);
-
-   // keep compiler happy
-   return true;
-}
-#endif
 
 namespace {
 
@@ -1385,8 +585,9 @@ SEXP rs_GEplayDisplayList()
 Error run(const ROptions& options, const RCallbacks& callbacks) 
 {   
    // copy options and callbacks
-   s_options = options ;
-   s_callbacks = callbacks ;
+   s_options = options;
+   s_callbacks = callbacks;
+   setStdCallbacks(&s_callbacks);
    
    // set to default "C" numeric locale as-per R embedding docs
    setlocale(LC_NUMERIC, "C") ;
@@ -1540,7 +741,7 @@ Error run(const ROptions& options, const RCallbacks& callbacks)
                             loadInitFile,
                             s_options.saveWorkspace,
                             cb,
-                            &s_internalCallbacks);
+                            stdInternalCallbacks());
 
    // keep compiler happy
    return Success() ;
@@ -1711,50 +912,13 @@ void setImageDirty(bool imageDirty)
    R_DirtyImage = imageDirty ? 1 : 0;
 }
 
-bool imageIsDirty()
-{
-   return R_DirtyImage != 0;
-}
-
 bool browserContextActive()
 {
    return Rf_countContexts(CTXT_BROWSER, 1) > 0;
 }
    
-void quit(bool saveWorkspace, int status)
-{
-   // invoke quit
-   std::string save = saveWorkspace ? "yes" : "no";
- #ifdef _WIN32
-   std::string quitErr;
-   bool didQuit = win32Quit(save, 0, true, &quitErr);
-   if (!didQuit)
-   {
-      REprintf((quitErr + "\n").c_str());
-      LOG_ERROR_MESSAGE(quitErr);
-   }
- #else
-   Error error = r::exec::RFunction("base:::q", save, status, true).call();
-   if (error)
-   {
-      REprintf((r::endUserErrorMessage(error) + "\n").c_str());
-      LOG_ERROR(error);
-   }
- #endif
-}
-   
 namespace utils {
    
-bool isR3()
-{
-   return s_isR3;
-}
-
-bool isR3_3()
-{
-   return s_isR3_3;
-}
-
 bool isPackratModeOn()
 {
    return !core::system::getenv("R_PACKRAT_MODE").empty();
@@ -1784,6 +948,11 @@ const FilePath& userHomePath()
    return s_options.userHomePath;
 }
 
+FilePath logPath()
+{
+   return s_options.logPath;
+}
+
 FilePath safeCurrentPath()
 {
    return FilePath::safeCurrentPath(userHomePath());
@@ -1807,16 +976,6 @@ FilePath tempDir()
       LOG_ERROR(error);
    FilePath filePath(r::util::fixPath(tempDir));
    return filePath;
-}
-
-SuppressOutputInScope::SuppressOutputInScope()
-{
-  s_suppressOutput = true;
-}
-
-SuppressOutputInScope::~SuppressOutputInScope()
-{
-   s_suppressOutput = false;
 }
 
 } // namespace utils
